Рефакторинг. Объединил простые интерфесы в один файл.
