--- SOLID pattern/Interfaces/I3DShape.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SOLID_pattern.Interfaces
-{
-    public interface I3DShape : ICalculateArea, ICalculateVolume
-    {
-    }
-}
