--- SOLID pattern/Interfaces/ICalculateArea.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SOLID_pattern.Interfaces
-{
-    public interface ICalculateArea
-    {
-        /// <summary>
-        /// Вычисление площади
-        /// </summary>
-        /// <returns></returns>
-        double CalculateArea();
-    }
-}
