--- SOLID pattern/Interfaces/ICalculatePerimeter.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SOLID_pattern.Interfaces
-{
-    public interface ICalculatePerimeter
-    {
-        /// <summary>
-        /// Вычесление периметра
-        /// </summary>
-        /// <returns></returns>
-        double CalculatePerimeter();
-    }
-}
