--- SOLID pattern/Interfaces/ICalculateVolume.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SOLID_pattern.Interfaces
-{
-    public interface ICalculateVolume
-    {
-        double CalculateVolume();
-    }
-}
